+++ CHANGELOG.md
@@ -11,6 +11,9 @@ _released `TBD`_
 - ❗️ BREAKING: prioritize `pytest` before Django's `./manage.py test`. `pytest` can run Django tests via [django-pytest](https://pytest-django.readthedocs.io/en/latest/), but not the other way around
 - ❗️ BREAKING: stop looking for the `tests.py` file, which isn't a standard. Instead if there are any python-related files (such as `pyproject.toml` or `.venv`), run `python -m unittest`
 - add better `pytest` detection in projects without a `.pytest-cache` (https://github.com/xavdid/universal-test-runner/pull/7)
+- add support for justfiles with non-default names:
+  - `Justfile`
+  - `.justfile`
 - flush the stream after sending the screen-clearing escape sequence, which should fix issues where buffered output from test runners is cleared when it shouldn't be
 - ❗ BREAKING: tweak how `UTR_CLEAR_PRE_RUN` and `UTR_DISABLE_ECHO` are read from the environment. Previously, _any_ set value would engage the option. Now, any value other than `0` is considered "present". So, if you were setting them to `0` before in the hopes of activating those options, set them to `1` instead